Create Implementation of Ml using AI
This project is about implementation of machine learning using Artificial intelligence. Here this model is design for the classification of image using Artificial intelligence.

The model is a Convolutional Neural Network (CNN) designed for image classification tasks. It takes an input image, processes it through multiple layers, and outputs a probability distribution over the possible classes.
Hence, it could classify any image in JPEG or PNG format.